src/app.py
# Main application file for image annotation app
import tkinter as tk
from tkinter import filedialog
from PIL import Image, ImageTk
import os

class AnnotationApp:
    def __init__(self, root):
        self.root = root
        self.root.title("Image Annotation App")
        # TODO: Implement GUI components

if __name__ == "__main__":
    root = tk.Tk()
    app = AnnotationApp(root)
    root.mainloop()