requirements.txt
opencv-python
Pillow
labelme